Added Order class
Added a base Order class and necessary Service/Dao classes.  The classes will be used to send orders to suppliers straight from the application, removing a product from the LowList if an order is open.

The bare basics are working now, still have to work it out further for it to do something meaningful.