Product deletion now sets a product as 'inactive'
Earlier when you deleted a product it would be gone from the database, causing errors when looking at the history of orders. Product now has an additional property, 'active', that prevents the product from displaying in any lists if set to false, while still keeping the entry in the database.